% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{events_to_regressors}
\alias{events_to_regressors}
\title{Generates regressors from a named list of Events}
\usage{
events_to_regressors(events, dates, h = NULL, seasonal = T,
  time_series = NULL, K = 3)
}
\arguments{
\item{events}{List of Event objects like the output from table_to_events().}

\item{dates}{Sequence of Date objects one for each row of the regressor table.}

\item{h}{Optional days to forecast for generating forecast regressors.}

\item{seasonal}{If true this will generate a single seasonal component with forecast::fourier(time_series, K=K) or
forecast::fourierf(time_series, K=K, h=h) if h was specified.}

\item{time_series}{ts object that is the data you are trying to forecast.}

\item{K}{input to forecast::fourier() and forecast::fourierf()}
}
\value{
Returns an xreg data.frame groups by event name.
}
\description{
Generates regressors from a named list of Events
}
\examples{
regs_table <- data.frame(event=c("christmas", "christmas", "promotion"),
                         start=c("2014-12-25", "2015-12-25", "2015-11-27"),
                         end=c("2014-12-25", "2015-12-25", "2015-11-30"),
                         time_function= c("none","none","inverse"),
                         stringsAsFactors=FALSE)

events_to_regressors(table_to_events(regs_table),
                     seq(as.Date("2015-11-01"), as.Date("2016-01-01"), by = 1))

}

