% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{table_to_events}
\alias{table_to_events}
\title{Generates list of Events from table.}
\usage{
table_to_events(regressors_table)
}
\arguments{
\item{regressors_table}{Data frame formatted as follows with all fields as strings.
\code{event      start       end         time_function}
\code{christmas  2014-12-25  2014-12-25  none         }
\code{christmas  2015-12-25  2015-12-25  none         }
\code{promotion  2015-11-27  2015-11-30  inverse      }}
}
\value{
Returns a list of Event reference classes.
    Grouping by event name.
}
\description{
Generates list of Events from table.
}
\examples{
regs_table <- data.frame(event=c("christmas", "christmas", "promotion"),
                         start=c("2014-12-25", "2015-12-25", "2015-11-27"),
                         end=c("2014-12-25", "2015-12-25", "2015-11-30"),
                         time_function= c("none","none","inverse"),
                         stringsAsFactors=FALSE)

table_to_events(regs_table)

}

